SOLR-15736: Split into own API file
